feat: move lr scheduler to training utils file
