+++ mava/systems/ff_ippo.py
@@ -44,13 +44,10 @@ from mava.types import (
 )
 from mava.utils import make_env as environments
 from mava.utils.checkpointing import Checkpointer
-from mava.utils.jax import (
-    make_learning_rate_schedule,
-    merge_leading_dims,
-    unreplicate_learner_state,
-)
+from mava.utils.jax import merge_leading_dims, unreplicate_learner_state
 from mava.utils.logger import LogEvent, MavaLogger
 from mava.utils.total_timestep_checker import check_total_timesteps
+from mava.utils.training import make_learning_rate_schedule
 
 
 def get_learner_fn(
