+++ mava/systems/rec_ippo.py
@@ -46,9 +46,10 @@ from mava.types import (
 )
 from mava.utils import make_env as environments
 from mava.utils.checkpointing import Checkpointer
-from mava.utils.jax import make_learning_rate_schedule, unreplicate_learner_state
+from mava.utils.jax import unreplicate_learner_state
 from mava.utils.logger import LogEvent, MavaLogger
 from mava.utils.total_timestep_checker import check_total_timesteps
+from mava.utils.training import make_learning_rate_schedule
 
 
 def get_learner_fn(
